fix: combine two file type to one
